형성평가2
Accepted(40)점으로 계산되어 다시 Success(100)점으로 고침

+++ 주석문과 디버깅/디버깅 형성평가2.c
@@ -5,7 +5,7 @@ int main() {
 	scanf("%lf %lf %lf", &a, &b, &c);
 	int sum = (int)a + (int)b + (int)c;
 	printf("sum %d \n", sum);
-	printf("avg %.0f \n", (a + b + c) / 3);
+	printf("avg %d \n", (int)(a + b + c) / 3);
 
 	return 0;
 }
